test(memory): cover main()'s real environ/home defaults, fix stale retro citation
No test called main() without explicit env/home overrides, so the two
default-resolution seams (os.environ if env is None else env; Path.home()
if home is None else home) could be swapped or broken with every existing
test still green. Adds one test per default, each isolating the seam it
proves (the other input supplied explicitly), with a negative control
confirming both fail when the defaults are inverted.

Adding those tests pushed test_importer_resolution.py to 533 lines,
regressing the taste-count-ratchet baseline the same way this PR's own
earlier rounds already hit (576 > 575). Splits TestImportMemoriesMain out
to tests/claude_mem/test_importer_main_integration.py, the same pattern
already used four times in this PR, restoring the count to baseline.

Also resyncs a stale line citation in the retrospective: the unqualified
command occurrences in MEMORY-MANAGEMENT.md are now at lines 167, 491, and
524, not 167, 487, and 520.

Refs #4780
