+++ tests/claude_mem/test_importer_resolution.py
@@ -4,6 +4,9 @@ Split out of tests/test_claude_mem_scripts.py, which covers the export scripts.
 The importer's resolution contract (precedence, blank handling, tilde expansion,
 exit codes) grew past the point where one module could hold both without
 breaching the 500-line taste-lint ceiling.
+
+The main() entry-point integration tests were split out again, to
+tests/claude_mem/test_importer_main_integration.py, for the same ceiling.
 """
 
 from __future__ import annotations
@@ -339,152 +342,3 @@ class TestVanishedImporterUsesStoredClassification:
 
         assert result == 1
 
-
-class TestImportMemoriesMain:
-    def test_exits_0_and_skips_when_optional_plugin_absent(self, tmp_path: Path, capsys) -> None:
-        result = _import_mem.main([], env={}, home=tmp_path / "empty-home")
-
-        assert result == 0
-        assert "SKIP" in capsys.readouterr().out
-
-    def test_exits_1_when_explicit_importer_does_not_exist(self, tmp_path: Path) -> None:
-        missing = tmp_path / "missing-importer.ts"
-
-        result = _import_mem.main(["--importer", str(missing)], env={}, home=tmp_path)
-
-        assert result == 1
-
-    def test_exits_1_when_configured_environment_importer_does_not_exist(
-        self, tmp_path: Path
-    ) -> None:
-        env = {_import_mem.IMPORTER_ENV_VAR: str(tmp_path / "missing-importer.ts")}
-
-        result = _import_mem.main([], env=env, home=tmp_path)
-
-        assert result == 1
-
-    def test_exits_1_when_configured_importer_is_a_directory(
-        self, tmp_path: Path, monkeypatch
-    ) -> None:
-        # exists() is true for a directory too. A misconfigured --importer (or
-        # an upstream marketplace layout change) that names a directory must
-        # be rejected by the guard, never handed to tsx as an argument.
-        importer_dir = tmp_path / "importer.ts"
-        importer_dir.mkdir()
-        memories = tmp_path / "memories"
-        memories.mkdir()
-        (memories / "shared.json").write_text("{}", encoding="utf-8")
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-        calls = []
-
-        def _record_call(*args, **kwargs):
-            calls.append(args)
-            return subprocess.CompletedProcess(args[0], 0, "", "")
-
-        monkeypatch.setattr(_import_mem.subprocess, "run", _record_call)
-
-        result = _import_mem.main(["--importer", str(importer_dir)], env={}, home=tmp_path)
-
-        assert result == 1
-        assert calls == []
-
-    def test_exits_1_when_configured_importer_fails(self, tmp_path: Path, monkeypatch) -> None:
-        importer = tmp_path / "importer.ts"
-        importer.write_text("// stub importer", encoding="utf-8")
-        memories = tmp_path / "memories"
-        memories.mkdir()
-        (memories / "shared.json").write_text("{}", encoding="utf-8")
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-        monkeypatch.setattr(
-            _import_mem.subprocess,
-            "run",
-            lambda *a, **kw: subprocess.CompletedProcess(a[0], 3, "", "boom"),
-        )
-
-        result = _import_mem.main(["--importer", str(importer)], env={}, home=tmp_path)
-
-        assert result == 1
-
-    def test_exits_0_when_configured_importer_succeeds(self, tmp_path: Path, monkeypatch) -> None:
-        importer = tmp_path / "importer.ts"
-        importer.write_text("// stub importer", encoding="utf-8")
-        memories = tmp_path / "memories"
-        memories.mkdir()
-        (memories / "shared.json").write_text("{}", encoding="utf-8")
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-        monkeypatch.setattr(
-            _import_mem.subprocess,
-            "run",
-            lambda *a, **kw: subprocess.CompletedProcess(a[0], 0, "", ""),
-        )
-
-        result = _import_mem.main(["--importer", str(importer)], env={}, home=tmp_path)
-
-        assert result == 0
-
-    def test_exits_0_when_no_memory_files_present(self, tmp_path: Path, monkeypatch) -> None:
-        importer = tmp_path / "importer.ts"
-        importer.write_text("// stub importer", encoding="utf-8")
-        memories = tmp_path / "memories"
-        memories.mkdir()
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-
-        result = _import_mem.main(["--importer", str(importer)], env={}, home=tmp_path)
-
-        assert result == 0
-
-    def test_creates_the_memories_dir_when_it_is_absent(self, tmp_path: Path, monkeypatch) -> None:
-        """The absent-directory branch, which the empty-directory case never reaches."""
-        importer = tmp_path / "importer.ts"
-        importer.write_text("// stub importer", encoding="utf-8")
-        memories = tmp_path / "memories"
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-        assert not memories.exists()
-
-        result = _import_mem.main(["--importer", str(importer)], env={}, home=tmp_path)
-
-        assert result == 0
-        assert memories.is_dir()
-
-    def test_creates_nested_memories_dir_when_parents_are_absent(
-        self, tmp_path: Path, monkeypatch
-    ) -> None:
-        """parents=True is load-bearing: the parent may not exist either."""
-        importer = tmp_path / "importer.ts"
-        importer.write_text("// stub importer", encoding="utf-8")
-        memories = tmp_path / "absent-parent" / "memories"
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-
-        result = _import_mem.main(["--importer", str(importer)], env={}, home=tmp_path)
-
-        assert result == 0
-        assert memories.is_dir()
-
-    @pytest.mark.parametrize("blank", ["", "   "], ids=["empty", "whitespace"])
-    def test_exits_1_when_explicit_argument_is_blank(self, blank: str, tmp_path: Path) -> None:
-        home = tmp_path / "home"
-        _make_claude_default(home)
-        env = {_import_mem.IMPORTER_ENV_VAR: str(_make_claude_default(tmp_path / "other"))}
-
-        result = _import_mem.main(["--importer", blank], env=env, home=home)
-
-        # A usable env value and a usable default both exist; the blank argument
-        # must still fail rather than fall through to either of them.
-        assert result == 1
-
-    def test_exits_1_when_importer_binary_is_unavailable(self, tmp_path: Path, monkeypatch) -> None:
-        importer = tmp_path / "importer.ts"
-        importer.write_text("// stub importer", encoding="utf-8")
-        memories = tmp_path / "memories"
-        memories.mkdir()
-        (memories / "shared.json").write_text("{}", encoding="utf-8")
-        monkeypatch.setattr(_import_mem, "_MEMORIES_DIR", memories)
-
-        def _raise(*_args, **_kwargs):
-            raise FileNotFoundError("npx not on PATH")
-
-        monkeypatch.setattr(_import_mem.subprocess, "run", _raise)
-
-        result = _import_mem.main(["--importer", str(importer)], env={}, home=tmp_path)
-
-        assert result == 1
